requirements.txt
aldryn-addons==1.0.0
bleach==1.4.2
dj-database-url==0.4.0
Django<1.9
django-admin-extend==0.0.3
django-auth-ldap==1.2.7
django-ckeditor==5.0.3
django-classy-tags==0.7.1
django-cms==3.2.1
django-easy-select2==1.3.2
django-formtools==1.0
djangocms-forms==0.1.16
django-getenv==1.3.1
django-reversion==1.8.7
django-sekizai==0.9.0
Django-Select2==4.3.2
django-taggit==0.18.0
django-treebeard==4.0
djangocms-admin-style==1.1.0
djangocms-column==1.6.0
djangocms-file==0.2.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.1.1
djangocms-installer==0.8.6
djangocms-link==1.7.1
djangocms-picture==0.2.0
djangocms-style==1.6.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.8.1
djangocms-video==0.2.0
django-cms-smartsnippets==0.1.3
GeoIP==1.3.2
html5lib==0.9999999
mysqlclient==1.3.7
# MySQL-python==1.2.5
Pillow==3.1.1
python-ldap==2.4.25
pytz==2015.7
requests==2.9.1  # check if we need this
six==1.10.0
tzlocal==1.2
bleach==1.4.2
